backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "theta-gainers-backend"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "uvloop>=0.19; sys_platform != 'win32'",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "redis[hiredis]>=5.0",
    "structlog>=24.1",
    "bcrypt>=4.1",
    "pyjwt>=2.8",
    "pyotp>=2.9",                   # TOTP 2FA
    "cryptography>=42",             # encrypting broker tokens at rest
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "orjson>=3.9",
    "tenacity>=8.2",                # retry with backoff
    "sentry-sdk[fastapi]>=1.40",
    "boto3>=1.34",                  # S3 / SES / Secrets Manager
    "apscheduler>=3.10",            # daily security master sync, EOD recon
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "httpx>=0.27",
    "black>=24.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "types-redis",
    "freezegun>=1.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "ASYNC", "S", "A", "C4", "T20", "SIM", "RET"]
ignore = ["S101"]  # allow asserts in tests

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra --cov=app --cov-report=term-missing --cov-fail-under=80"
